Merged PR 5309: Return value according to whether the resubscribe was successful
Resubscribe messages get 'stuck' in the importer table and are getting processed every importer cron run. The status of the record remains not_imported and never gets set to success/failure.

####The fix
Return true/false. The return of this method gets checked by isset an if so, the importer record is marked as success/failed depending on the result.

If isset is false, then the importer record remains as NOT_IMPORTED and gets picked up again the next time the importer cron runs.

####Steps to Reproduce
Unsubscribe a magento customer in dotmailer
Sync
Resubscribe in Magento
Sync importer

#####Actual Result
Importer record remains 'Not Processed' despite being processed

#####Expected Result
Importer record gets set to Processed or Failed so that it's not picked up again

Related work items: #62545