Merged PR 5366: fix the locale for order purchase date
The Problem
Order created date(purchaseDate) is sending the date that is used in default config for the timezone.

The fix
Time is stored in UTC for the orders and there is no need for localization because dotmailer is keeping the dates in the same UTC format.

Also there is no "purchase_date" in M2, we are senging only "purchaseDate" field for orders and this is somehow transformed to match the shema.

Related work items: #58400