Merged PR 5901: Pass required parameters to dotmailer api method
As content is being run through the personalizer, we have to specify an email address to personalise for. It doesn't matter that it doesn't exist. It won't be used.

As content is being compiled and personalized, no need to further convert content. Resized images that were on vedimage will now be on the cdn. Unsub links will be removed. Forward to a friend and cant read will remain, but i think this is ok. We caveat the docs for smtp and api method to say that they won't be replaced or valid. str_replace of "Unsubscribe", "Forward this email" and "View in browser" seems heavy handed as they may want to use these words/phrases in the email.

Related work items: #64426