Merged PR 6002: Fix for the default and website using the same template code
fix for the default and website using the same template code

When mapping the templates using the default and website level the store id may have the same ID which also defaults to 0.

Now we should use the default, website or store code when it's present - template name should end with the code depending on what level was mapped.

Related work items: #65245