Merged PR 6055: Merge calin/email-templates-iheritance to develop
When syncing template the default store/website is still using the ID for the code name and the wrong template will be synced.

This should replace the store ID with 'admin' code name.
Ignore Inherited options for the template if was not selected. By default we removed the ignore for inherited items so it will not be saved against the store but the parent option. Now we put back the check for inhereted items.

Related work items: #65352