Merged PR 6776: existing cart is never sent
existing cart is never sent

Existing cart for customer and guest is never sent.
After the last refactor the abandoned cart check is inluded to check if the abandoned cart items changed for the 2'nd and 3'rd and ignore the send. Where this may be a valid option for the AC1 the contents for the AC2 and AC3 will never change and should procced with the sending.

This should fix the shouldnot send for existing carts.

Fixed also the test that was renamed and test were failing.

Related work items: #67538