Merged PR 6783: change wishilst save observer to add product
change wishilst save observer to add product

Wishist is not triggering the save envent when the product is added to the product in magento 2.2.4 version and we should chnage the event observer to the add product one.

Switch to event to the wishlist_add_product when checking the if the wishlist is new.

Related work items: #67569, #67577