Merged PR 6862: update the ip for trial from the server
update the ip for trial from the server

We can send the server IP for the trial account so in case the client's IP in private network the trial account will be marked by security if is not public.

Check if HTTP_X_FORWARDED_FOR header is set and get ip from the list.

Related work items: #67745