Merged PR 7637: Consent data for the contacts.
Consent data for subscriber that will include the Text, Url, IP, Datetime, User agent.
Text  - should keep the informatino that is selected in config for the consent that is available for the subscribers and customers.
Url - is the url when the subscriber clicked to subscribed.
IP - is the subscriber IP address.
User agent - is the subscriber user agent.

Installation
When installing first time the consent should create new table or drop and recreate "email_contact_consent".
Upgrading
when upgrading the new table will be created "email_contact_consent" with following fields :

- id
- email_contact_id
- consent_url
- consent_datetime
- consent_ip
- consent_user_agent

The consent data will be created and stored into new table once when subscriber becomes subscribed and it will be removed as soon as the data was imported (marked with finished).

New section in admin will be created in Configuration -> Consent will hold the options
- customers or guest that clicked subscribed fom the customer registration or checkout page
- newsletter subscribscription from the footer or any other place

The data will be created and stored against the email consent table when the subscriber will become Subscribed or need to confirm is enabled. If need to confirm is enabled will put the consent data into the table but not sync so we know the data from the moment the subscriber clicked first time but not from the subscription confirmation.

Subscribers will be using the origin importer for subscribers with additional fields that will be included when this feature is enabled to use. Once the import will be created the consent data will wait until the import for the contact is marked Finished by the importer and the consent record will be removed in bulk.

How to test :
Enable one of the consent option and enter the consent data that will be stored agains the contact. Once the subscriber will be marked "Subscribed" or "Not Active" the consent data will be available in our table. Leave the normal subscriber sync to run and the point the csv file is created there should be consent data included into csv(if is not included check the subscribers with sales data csv or report a bug).

Once the import is finished the consent is removed from the table and cannot be reimported again.

At the current stage the consent data is not working in dotmailer and will assume that the data is created in account!

But sure you can find more different scenarious how to created and subscribe this is one simple scenario for now.

Related work items: #66017