Merged PR 9079: first and last categories for customers and subscribers with sales data
Customer and guest subscribers with sales data should have category names from the first and last order.

How to test
For Customers
Create a customer with multiple orders, run the sync and check import data or check the contact inside the account. The contact should contain the full category tree inside the "FIRST_CATEGORY_PUR" and "LAST_CATEGORY_PUR"

For Subscribers
Create a guest subscriber and confirm as a subscriber if needed then place some orders using the email for subscriber as a guest on checkout. Enable sync subscribers with sales data in configuration settings. Run the subscriber sync and check the csv file or contact in the account should contain the full category tree inside the "FIRST_CATEGORY_PUR" and "LAST_CATEGORY_PUR".

Related work items: #72729, #72731