Merged PR 11275: Handle font families correctly
Escape font family correctly for coupon code dynamic content styling.

Previously, when setting the font in teh configuration setting to a font with a space in it (and therefore surrounding apostrophes) the styling would break completely. This was due to the whole font-family being escaped and changing ' to appropriate characters. Now we split teh font-family on ',', loop over each font and escape it and rebuild the surrounding apostrophes if need be.

Iterate over each font, if no apostrophe, then escape the font name, if there is apostrophes, strip them, escape the font, add apostrophes back in and return.

####How to test
In Magento go to Stores -> Configuration -> dotdigital -> Configuraiton -> Dynamic Content Styling
Change Coupon Code Font to be a one with a space in it
View the coupon code link and ensure the font is applied correctly

Check that other fonts, without a space continue to function as expected.

Related work items: #81220