Merged PR 12348: EDC for abandoned cart changed
EDC for abandond cart changed

If we add a configurable product into basket (see img 1)
![img1.png](https://dev.azure.com/dotdigital/b2ba4bc1-4a9a-48f7-a69a-5cf84134b7c4/_apis/git/repositories/cc22fc8a-89ce-4edf-ba9c-81a01b55e11a/pullRequests/12348/attachments/img1.png)

In edc we need to show the same image as the product we selected.

But currently we are showing the parent product image in EDC. (see img2)
![img2.png](https://dev.azure.com/dotdigital/b2ba4bc1-4a9a-48f7-a69a-5cf84134b7c4/_apis/git/repositories/cc22fc8a-89ce-4edf-ba9c-81a01b55e11a/pullRequests/12348/attachments/img2.png)

In EDC We should show this (img3)
![img3.png](https://dev.azure.com/dotdigital/b2ba4bc1-4a9a-48f7-a69a-5cf84134b7c4/_apis/git/repositories/cc22fc8a-89ce-4edf-ba9c-81a01b55e11a/pullRequests/12348/attachments/img3.png)

So for configurable products, we must show in EDC, parent product Title, Parent Product URL,Child(selected) product Image , Child (selected) Product Price.

For a Simple visible product, we leave the logic as it is.

For grouped products, we need to show all simple products that being a part of this separetelty,

For Bundled Products, only the parent product!

Related work items: #80913