Merged PR 13526: Remove escapeHtmlAttr and hard code attribute
`escapeHtmlAttr()` was only introduced as a method in blocks from 2.2. Prior to 2.2 it was possible to use the `\Magento\Framework\Escaper` directly via a helper, but after 2.1 this is not possible. Since it's not possible to have a single, safe escaping method here that is compatible with 2.1-2.3+, it's better just to hard code the title text.

Related work items: #88155