Merged PR 13649: Resubscribe recently resubscribed users in EC, who subscribe in between syncs
A timing condition can result in a user being suppressed in EC, then resubscribe in Magento before the Subscriber sync runs and marks them as suppressed in Magento. As a result of this, they are not resubscribed in EC by a subsequent sync.

To test: follow steps as per ticket #88350

- subscribe a new email address in Magento
- Sync subscriber , sync importer.
- Suppress this email in EC
- Create an new account with same email address in Magento and Sign up for Newsletter
- Sync contact, sync subscribers, Run importer twice (!)

On the second importer run, the subscriber will be resubscribed in EC.