Merged PR 13665: Filter processed quotes with active check
Both `getStoreQuotes()` and `updateDataFieldAndCreateAc()` check for active quote status, but only when processing the first abandoned cart campaign in the series. Thereafter we get the quote ids from our own `email_abandoned_cart` table and assume that they are still valid. This PR adds a filter for `is_active` on the quote collection for numbers 2 and 3 in the series.

Related work items: #87961