Merged PR 14727: Check if file exists before cleaning and archiving imported CSV files
## What's being changed

Processing of finished importer items.

## Why it's being changed

It is theoretically possible for files stored in `var/export/email` and `var/export/email/archive` to be removed after their import has started, but before their import has been marked as finished. This is the unlikely snafu responsible for the bug in #64134. Here we confirm that the file actually exists before continuing, lest an exception is thrown when either cleaning or archiving (which could break the importer).

We check both the email folder and the archive folder because cleaning evidently happens on files in either location.

## How to review / test this change

- Set up a few pending imports for either Contacts, Subscribers or Guests (any import involving CSV transfer)
- Run the importer (this will put the status to 1 i.e. importing)
- Remove the related .csv file from the email folder
- Re-run the importer
- Confirm no exception is thrown if the file is not found
- Confirm subsequent importer files are processed/archived as normal

Related work items: #64134