Merged PR 14926: Add an optional version number to tracking script query string
This PR allows an optional version number to be appended to the `_dmpt` and `_dmmpt` tracking scripts, to force a refresh in a users' browsers (as the file would be cached locally and by Cloudflare).

To use this, a config item `connector_configuration/tracking/script_version` can be added to Magento, followed by flushing the cache.

Additionally, the ROI tracking code send has been wrapped in a jQuery document.ready event to defer execution until we're sure that the `window._dmTrack` function is available.