Merged PR 14863: Use setFromAddress to display friendly from name on transactional
## What's being changed

Uses `$message->setFromAddress()` in place of `$message->setFrom()`, conditional on Magento core support, in the Message Plugin.

## Why it's being changed

After Zend Framework 1 was retired in 2.2.8, `$message->setFrom()` set the from email address but did not set the 'Friendly from name' set in Engagement Cloud for synced campaign templates. In 2.2.9, `$message->setFromAddress()` was introduced to fix this.

| Version | Framework | Method |
|-----------|:-----------:|-----------:|
| 2.2 > 2.2.7 | \Zend\Mail\Message | setFrom($fromAddress, $fromName) |
| 2.2.8 | \Magento\Framework\Mail\Message | setFrom($fromAddress) |
| 2.2.9 + | \Magento\Framework\Mail\Message | setFromAddress($fromAddress, $fromName = null) |

Note that `setFrom()` up to 2.2.7 did take `$fromAddress` and `$fromName` params thanks to `Zend_Mail`. So with this change, it is only 2.2.8 that will not support friendly from names in transactional emails.

Note that there remains a problem with 2.1, related to the `TemplatePlugin`.

## How to review / test this change

- Activate DDG SMTP
- Set up a sent campaign in EC with a friendly from name
- Configure a transactional send to use an EC template e.g. newsletter subscribe, wishlist share
- Trigger the transactional send
- Download the .eml file from EC transactional outbox
- Confirm that there is a friendly from name
-- e.g. From: "Chaz" <Dotmailer53@r1.dotmailer-email.com> [2.2+]
-- e.g. From: "Owner" <Dotmailer53@r1.dotmailer-email.com> [2.1]
- All versions should pass, except 2.2.8

## Notes

- Must test in 2.2 (< 2.2.8), 2.2.8, and 2.3

Related work items: #92606