Merged PR 15776: Add 204 response code if no content found
Actual problem:

Magento M2 extension, EDC When the EDC returns an empty page, it still returns a 200 response, see here: https://www.screencast.com/t/NaA0Mol7T This means the campaign will get sent, but the EDC block will be empty.

Expected result:

How external dynamic content URLs which don't return data are handled

If a request for external dynamic content for a contact returns no content (an HTTP 204 status code), the contact is skipped but sending to the rest of the contacts continues. This is recorded and your report overview for the campaign will contain an 'Email skipped' section.

**How to test**
Checkout to this branch and **UPGRADE AND RECOMPILE**
Check all the edc's in your browser by adding an INVALID quote_id or customer_id or any other parameter that EDC Url Requires

Open the network tab. Your network status code should be 204 instead of 200 it was before

Related work items: #93925