Merged PR 16294: Add method to UrlFinder to retrieve a product's small image
## What's being changed

This PR adds a method to the `UrlFinder` class to retrieve a product's `small_image`. If the product does not have an image, its parent's will be returned.

## Why it's being changed

A parent product image was not being added to synced product data if a simple product did not have an image.

## How to review / test this change

- Remove all images from a simple product with a parent configurable product
- Run a catalog sync and check that the simple product has been assigned it's parent's image

## Notes

The previous `getProductImageUrl()` method in `UrlFinder` did not appear to work in the context of syncing Catalog products, but will still be used for EDC templates. Testing should include EDC templates.

Related work items: #97385