Merged PR 16293: Use SenderResolverInterface to set From values
## What's being changed

We are adding our own concrete implementation of the Sender Resolver interface in order to set stored From email and From name values.

## Why it's being changed

Up until now, we have used our `MessagePlugin` to intercept the mail transport process and update the current From name and From email from the `email_template` table. However, in Magento 2.3.3 the method we are intercepting is no longer used, so we needed to take a different approach.

The `SenderResolver` implementation is in fact better anyway because:
a) it's more specific than setting From values while setting Body content and
b) it allows us to remove some of the cumbersome `instanceof \Zend_Mail` checks we were forced to use to handle 2.1 and 2.2+

## How to review / test this change

- Map templates to various transactional emails - pick New Order, Newsletter Success, New Account (these all use different core Magento modules so it's important to test all of them).
- Set up DDG SMTP
- Trigger the sends
- Ensure Friendly from name and email address are set correctly as per the relevant template's stored values
- Test in a different store scope, with an alternate template mapping to the default level.

### Notes

- Must test in 2.1, 2.2, 2.2.8, 2.3.
- Re-check encoding - use umlauts in subject, body and from name values.

Related work items: #97408