Merged PR 17144: Remove empty block in favour of virtual type
## What's being changed

We're removing an empty block class and instead using di.xml to define a virtual type.

## Why it's being changed

This is a suggestion from Magento core team.

## How to review / test this change

- Clear cache
- Enable chat and check the widget loads on the front end

## Notes

The virtual type is namespaced with the term 'Virtual' in order to indicate that this class doesn't actually exist in a concrete form in our extension code. See https://www.yireo.com/blog/2018-05-06-magento-2-virtual-types-with-namespaces.

Related work items: #100438