Merged PR 17179: Ensure EC contact is created before posting cartInsight data
## What's being changed

The order of method calls when sending cartInsight data during the abandoned cart cron.

## Why it's being changed

We were sending cartInsight data before we had confirmed the presence of a contact on EC. This meant that in some setups, cartInsight data was not sent, and lost basket emails had missing content.

## How to review / test this change

*Email series*
- Drop a cart as a guest
- Run the abandoned cart sync without running any other syncs beforehand
- Confirm contact created in EC
- Confirm cartInsight data posted to new contact

*Program enrolment*
- Register as a new customer
- Drop a cart
- Run the abandoned cart sync without running any other syncs beforehand
- Confirm contact created in EC
- Confirm cartInsight data posted to new contact

Related work items: #100620