Merged PR 17609: Skip orders with no products
## What's being changed

Order sync - specifically the step at the end of the sync, where we mark all ordered products as unprocessed in order to mark them for the next catalog sync.

## Why it's being changed

We found a bug where apparently in some versions of Magento it's possible for orders with certain statuses to have no products (or at least to return none when calling `$order->getProducts()`).

## How to review / test this change

- Go to 224 test environment
- Hit Developer > Reset Order Sync
- The click Order sync > Run now
- The sync should complete as intended

Related work items: #101824