Merged PR 17460: Escape exclusion rule condition values on front end
## What's being changed

The controller that handles conditions attached to exclusion rules.

## Why it's being changed

Magento have suggested an improvement to our earlier fix for an identified XSS vulnerability in this area of the admin. This PR escapes submitted form values as opposed to values about to be stored in the database.

## How to review / test this change

Follow the replication steps on the [original ticket](https://dev.azure.com/dotdigital/ec/_git/ec-core-magento2-extension/pullrequest/16771):
- Click on Marketing->Exclusion Rules and then click on "Add New Rule"
- Enter arbitrary non-ambiguous rulename, and select arbitrary website
- Click on the Conditions tab, and click on "Add New Condition"
- Select Attribute to "Email", Condition to "is" (arbitrary) and Value to "" (leave empty, arbitrary).
- Now you have to edit the Form which is sent to the server:
For each of the three `"<input>"/"<select>"` fields (Attribute, Condition and Value)
Rightclick -> inspect element, and modify the "name" field of the HTML element
from something like `condition[_1562168869914_914][attribute]` to `condition['><script>alert(1)</script><i a='][attribute]`. (leave the second array key of condition as it is!)
- Now click on save, you should retrieve a "successfully saved" message
- As soon as someone clicks on "Edit" next to the newly created url, the XSS is triggered.
- Confirm that no alert appears.

Related work items: #101938