Merged PR 17781: Check data field name length before creating via API
## What's being changed

We're improving how we handle the creation of new data fields in Data Mapping.

## Why it's being changed

This PR adds a length check (in line with the limit set in Engagement Cloud for 20 characters) on all new data field names, before posting data via the API. Before this change, we posted the data and allowed EC to truncate the name silently, without informing the merchant.

## How to review / test this change

- Go to Engagement Cloud > Data Mapping > Create New Data Field
- Create a data field with a long name like CHAZ_KANGEROO_HOODIE_PREFERENCE
- See an error message above the form
- Do not see the long data field in EC > Contacts > Contact Data Fields
- Double check regular data field names pass validation and are created in EC

Related work items: #102462