Merged PR 17849: Fix missing message attribute
**Steps to Reproduce**
Create a suppressed contact in Engagement Cloud.
Trigger an AC for that contact and a subscribed contact.
AC is processed ok, send records are created in magento
Then when campaign send cron picks up the 2 sends

**Actual Result**
Both fail due to one of them being suppressed

**Expected Result**
Magento knows the contact is suppressed and there is some logic to try and handle this. But it doesn't do what it should do and stop them being included in the send.

**What Changed**
Now we're looking for $response->status instead of $response->message in order
to handle the suppress contacts, as it looks like $response->message is not available any more.

https://dotmailer.zendesk.com/agent/tickets/436949
https://dotmailer.zendesk.com/agent/tickets/515984

Related work items: #80802