Merged PR 18942: Replace Json class with core SerializerInterface
## What's being changed

We're removing our Json serializer class and reverting to using the core class.

## Why it's being changed

`Magento\Framework\Serialize\SerializerInterface` was introduced in 2.2, and our `Model\Config\Json` serializer class was copied from core in order to be backwards compatible with 2.1. We've no need of this now that we no longer support 2.1.

There was some error handling code in the Importer class which could not have worked under any circumstances, so this has been replaced with exception handling.

## How to review / test this change

- Check all areas of functionality affected by the modified classes.
- Run any sync in order register an import queue item.
- Try to force the serializer to throw an exception.

Related work items: #106566