Merged PR 19609: Allow additional import types for bulk importer syncs
Process additionalImportTypes for bulk importer syncs

## What's being changed

The ImporterQueueManager.

## Why it's being changed

We need to be able to append additional import types to the importer queue. Specifically, in this case we want to allow a plugin in the B2B module to add a 'B2B_Quotes' type to be processed using the regular Bulk transactional data import logic.

## How to review / test this change

- On a CE Magento (i.e. no B2B), setup some pending imports and run the importer sync
- It should process and complete as normal

Related work items: #109323