Merged PR 20174: Add system messages to highlight cron errors
## What's being changed

This feature will check the `cron_schedule` table every 15 minutes, and show a system message if jobs with errors are found.

We are changing the default longevity of our cron group's failure history to 72 hours, in line with Magento's default settings, in order to preserve error messages for longer.

## Why it's being changed

This provides an additional layer of visibility to the merchant, specifically for dotdigital cron jobs.

## How to review / test this change

- Set up your crons to run
- After clearing cache you should start to see `ddg_automation_cron_monitor` jobs logged to the `cron_schedule`
- Set a few of the `ddg_automation_` jobs in the table to have the status 'error'
- After the cron runs again, refresh any page in the Magento admin to reveal a new system message indicating the cron jobs that have failed. You should also see a flag in the `flag` table.
- Check the link to the Cron Tasks report.
- Clear the errors and confirm the message (and the flag entry) goes away.

Related work items: #110531