Merged PR 20408: Set all references to transactional sync limit to default scope
## What's being changed

When we retrieve the config value for Transactional Data Sync Limit.

## Why it's being changed

Transactional Data Sync Limit was changed to have Global scope in 3.2.0. In catalog sync, we retrieve the value in the correct way, but we did not change Wishlist, Review and Order sync. In these syncs the key was still being referenced in 'website' scope, and called for each website in the loop.

## How to review / test this change

- Set a limit in Developer > Transactional Data Sync Limit
- Run Order, Review and Wishlist syncs and ensure the set value is respected

Related work items: #111124