Merged PR 20701: Fix data migration for split databases
## What's being changed

The part of our data migration process that takes records from `sales_order` and inserts them into our `email_order` table.

## Why it's being changed

The part was failing for split database set ups, where tables for quotes and sales and parcelled out to separate databases, to improve performance. We were halfway there, but we needed to explicitly define the database name in the select statement. If this is not set, the script will look for `sales_order` in the main database, and will fail.

## How to review / test this change

- Run `bin/magento dotdigital:migrate` in a split database (Commerce) Magento
- Run `bin/magento dotdigital:migrate` in a non-split database (Commerce) Magento
- Run `bin/magento dotdigital:migrate` in a non-split (Open Source) Magento

Related work items: #112674