Merged PR 21110: Remove logging in Email module when Chat is initialised
## What's being changed

We're removing the log line for 'Initialised for Chat' in the Email module. This is being moved to the Chat module.

## Why it's being changed

The `setUpChatAccount()` client method is now used when a) initialising chat when the settings are saved and b) whenever a token needs refreshing. So the logging needs to change to reflect the different actions.

## How to review / test this change

- Save chat settings in Magento admin
- Confirm there is no log line output

Related work items: #114450