Merged PR 21571: Return messages for crons listed in developer section
###What's being changed
We now return a sync result for all syncs that can be manually run through Dotdigital > Developer section.

###Why it's being changed
This fixes a regression introduced in 4.5.2. We recently removed these returns, but as a result we lost the printed success messages for some of these syncs, when they are triggered manually.

###How to review / test this change
- Go to Stores > Configuration > Dotdigital > developer
- Run manually all the syncs
- Make sure that messages are printed as expected

Related work items: #116024