Merged PR 21407: Send subscriber status data field in all subscriber syncs
## What's being changed

We're adding the SUBSCRIBER_STATUS data field to subscriber sync (with and without sales data). Currently this data field is only synced when the account becomes a customer.

## Why it's being changed

Evidently this data field is being used as an indicator or segment determiner in EC, and it's useful to have in all cases.

## How to review / test this change

- Subscribe to the newsletter
- Run subscriber sync
- See SUBSCRIBER_STATUS = 'Subscribed' in your csv
- Place an order
- Run subscriber sync
- See SUBSCRIBER_STATUS = 'Subscribed' in your csv alongside sales data
- Run contact sync
- Ensure the same account is synced again with the correct SUBSCRIBER_STATUS value
- Unsubscribe
- Run importer sync
- See that the contact has 'Unsubscribed' in the data field in EC

Related work items: #115714