Merged PR 22167: Handle empty subscriber status in subscriber export
## What's being changed

We're throwing an exception if we try to get a subscriber status string (like 'Subscribed' or 'Unsubscribed') by passing either a NULL value, or a value that doesn't exist in our prescribed list of possible statuses.

As well as fixing the specific 'Notice: undefined index on L708 of ContactData' bug, I've done a small tidy up of the `Subscriber` class to improve readability.

## Why it's being changed

In 4.5.3 we added `subscriber_status` data to the regular subscriber export: https://github.com/dotmailer/dotmailer-magento2-extension/commit/8e765d362d4b297751e4b985364aa11cabd344c9. However, in the case of the `SubscriberExporter` class, it seems that sometimes, the magic method `getSubscriberStatus()` called on an entity of type `\Dotdigitalgroup\Email\Model\Contact`, does not return the expected result.

Possible reasons for this:
- the status itself has changed in the space of time between making the initial `getSubscribersToImport()` query and the later call to `getSubscribersToImportFromEmails()`. This is quite unlikely, and we've not replicated this in testing.
- the magic method is not available at the time for some reason

So in all cases we will send up an empty string if we can't get a status; and in the specific case of `SubscriberExporter`, we're adding some additional logging to try and understand how this happens.

We can't be 100% sure that the magic method was at fault, but we've replaced it anyway with `getData('subscriber_status')`. This has the advantage of indicating that we're getting a property of the object as opposed to calling a local `getSubscriberStatus()` method (which we do, in other cases).

## How to review / test this change

- Ensure that in your email_contact table you have 1) subscribers who are customers 2) subscribers who are guests with no order history and 3) subscribers who are guests with order history
- Run subscriber sync
- Pass in a NULL status code to `getSubscriberStatusString($statusCode)` and check the exception
- Check all the other try / catches and make sure the fall back is an empty string

Related work items: #118055