Merged PR 24218: Catch unserialize exception when processing order items
## What's being changed

This is a workaround for possible unserialize errors encountered when processing order items.

## Why it's being changed

Evidently it's possible to have `product_options` stored against an order item as string, as opposed to JSON. If this is the case, and Magento can't unserialize it ([here](https://github.com/magento/magento2/blob/2.3.5/app/code/Magento/Sales/Model/Order/Item.php#L479)), the order sync fails.

This happens regardless of whether the setting to include product custom options in the order sync is selected or not; the exception is thrown in the initial loop to gather order items from `sales_order_item`.

This does mean that, if Magento can't unserialize the data for any one of the items in an order, no product data will be synced for that order. However, it seems like a rare situation.

## How to review / test this change

- Place an order
- Hack the `product_options` value in one of the order item rows relating to your order. Make it a string that can't be unserialized!
- Run order sync
- Observe log lines in connector.log
- Check the output in email_importer
- Repair your hack
- Re-run the sync as normal
- Confirm product data is restored to the import data

Related work items: #124795