Merged PR 24547: Check for an array before looping over order status automations
## What's being changed

We're adding an extra array check before looping over unserialized order status automation data.

## Why it's being changed

If the value stored in `connector_automation/order_status_automation/program` is `null`, try/catching the `unserialize` call handles this. However it is apparently possible for situations where there is an unserializable value, but the result is not an array. It's not clear how this happens - possibly legacy data. However, to avoid any invalid argument errors we are adding an additional array check.

## How to review / test this change

- Set the value of `$configStatusAutomationMap` to a string on L312
- Run the code and check it returns nothing and continues

Related work items: #126064