Merged PR 27060: Ensure failed cURL requests always return an object
## What's being changed

If the initial response from `curl_exec()` is falsey, we will now always return an object, and assign to it the curl error message.

## Why it's being changed

Our standard check for API request errors is `isset($response->message)`, because this is how errors are communicated back from the Engagement Cloud API. However there may be times where the `$response` has been false or null, and if we try to get e.g. `$response->status` merchants will see an error. See:
https://github.com/dotmailer/dotmailer-magento2-extension/issues/570.

## How to review / test this change

- Change your api endpoint in core_config_data to an invalid url
- Run Importer
- Make sure that execution completes as expected

Related work items: #133178