Merged PR 26694: Remove single update sync for wishlists
## What's being changed

We've refactored wishlist sync to remove the 'single' loop.

## Why it's being changed

Up until now the wishlist sync has run a 'bulk' loop, followed by a 'single' loop that picks up modified wishlists to be queued as either single updates or single deletes in the importer.

This PR moves all updates into the bulk update, i.e. all non-imported or modified wishlists are processed in one loop and queued together in one batch. If the core wishlist has no items, we queue a deletion in the importer.

The driver behind this was to resolve missing wishlist data in EC, which we fixed for bulk updates already, but not for single updates. An added benefit is that we now use the `updated_at` date in our own wishlist table instead of Magento's (which does not change when items are added or removed). We no longer set the `updated_at` date in `email_wishlist` when marking rows as imported.

## How to review / test this change

- Set up some wishlists as follows
- Website 1
  - Customer 1: new wishlist
  - Customer 2: existing wishlist, modified
  - Customer 3: existing wishlist, no items
- Website 2
  - Customer 4: new wishlist
- Run wishlist sync
- Check the importer table. Do the Bulk rows contain the expected data? Are Single Delete rows queued as expected?

**Schema modification**
- Prep your `email_wishlist` table by marking some wishlists as modified = 1
- Run `setup:upgrade`
- Check the column is now marked as 'deprecated'
- Confirm that the modified wishlists are now imported = 0 modified = NULL
- Add items to a wishlist (check table)
- Remove items from a wishlist (check table)
- Reset wishlists from admin (check table)

Related work items: #129372, #133869