Merged PR 27696: Enable resendable EDC coupons for recurring campaigns and programs
## What's being changed

In this PR we updated the coupon generation rule cases.

## Why it's being changed

This change supports merchants who enrol customers in recurring campaigns or programs, for example a birthday program that sends a coupon once a year. For various reasons the merchant may prefer to put an expiry on the coupon via the EDC param, yet nonetheless want the offer to renew each year without having to intervene.

Before this change, the email would go out to the customer on their birthday, but the send would be cancelled if they used the coupon last time around.

The new logic goes like this:
- If `code_allow_resend` is 1 and `code_cancel_send` is 1, and the coupon has expired
-- > Regenerate the coupon
- If `code_allow_resend` is 1 and `code_cancel_send` is 0, and the previous coupon has expired OR been used at all
-- > Regenerate the coupon

## How to review / test this change

- Create a new price rule and set an expiration date
- While generating ddg coupons for this rule confirm that this pattern works

```
If the Allow_resend is true and cancel_if_used is true,
We should generate a new one, when the previous coupon is expired.

If the Allow_resend is true and cancel_if_used is false,
We should generate a new one, when the previous coupon is expired OR if the coupon has already been used.
```

Related work items: #135728