Merged PR 28192: Check first if coupon is expired and update message
## What's being changed

In this PR we first check if the coupon is expired and then if it's used, in case that cancel send is set to yes.

## Why it's being changed

Because if cancel if used is YES, it will keep sending 204 if you've used a coupon in the past (that has now expired).

## How to review / test this change

- Create a cart rule, set cancel send to yes and generate a coupon code
- Use the coupon
- Regenerate the coupon (after the previous one expired)
- Make sure that you get a new coupon instead of 204

Related work items: #137880, #139032