Merged PR 27603: Update prices to reflect tax where required
## What's being changed

- Catalog insight
- Cart insight
- Web insight

## Why it's being changed

We have added new fields to these schemas for prices including tax. Separating prices excluding tax from prices including tax will in due course allow merchants to decide whether or not to include tax in their campaign templates and the abandoned cart block in Engagement Cloud.

All existing price fields are unchanged, with the exception of `totalPrice` in cart insight. This is changing - currently it _includes_ tax (if tax is applicable in the specific quote), but now it will _exclude_ tax.

For cart insight and web insight, prices including tax are what the customer sees on the website, if tax is applicable. Note that for guests, where a shipping address is not provided, tax will not be displayed and nor will it be captured (unless there is a default tax calculation, see below).

It is possible for a merchant to configure a default tax rate, and to set a default tax country (in Stores > Configuration > Sales > Tax > Default Tax Destination Calculation > Default Country) such that a base rate of tax is applied for all customers and guests. It is this rate that we reflect in catalog insight data.

## How to review / test this change

- see #27397, #27240, #27626

Related work items: #132445, #136042, #136750