Merged PR 28751: Enable PWA support for cart and review links in EDC
## What's being changed

In this PR we added an extra configuration field to allow merchants to define a separate non-Magento storefront url. This is then used to re-route the 'Take me to my basket' and 'Leave a review' link in the basket EDC.

## Why it's being changed

For merchants operating a PWA storefront, the basket EDC no longer worked because the 'Take me to basket' and 'Leave a review' link only redirected to a Magento frontend base URL path. With this change it is possible for the EDC to take customers back to the true storefront.

## How to review / test this change

**Abandoned Carts**
- Edit the PWA URL in Developer > PWA Settings. This should be an absolute path to your pwa instance (e.g https://pwa.local) (1)
- Set the Cart URL in Configuration > Abandoned Cart to whatever your storefront uses (e.g. /cart)
- Drop your cart
- Now run the AC EDC and click "Take me to my basket". You should be redirected to the checkout page.

**Reviews**
- Make sure that reviews edc url works as expected if settings are set as per (1)
## Notes

- Merchants using this configuration will want to also configure the 'Cart URL' (in Configuration > Abandoned Carts) to append to the `pwa_url`.
- With this option, we are unable to check whether the customer is logged in or not. So with PWA we can't go to the login page, then redirect again to the basket. So we're relying on the cart persisting on the customer's side, from the time they dropped the cart.

Related work items: #140469