Merged PR 29568: Offset cron job scheduling
## What's being changed

In this PR we added backend models to randomly set offsets for certain cron jobs. Currently our cron pattern is the following:
```
*/x * * * *
```
which means the cron runs every xth minute. For example: `*/30 * * * *` runs at 30 and 00 mins.

Now this is changed to:
```
x-59/y * * * *
```
which means that cron must run every x mins after the y'th minute (y generated randomly). For example `22-59/30 * * * *` runs at 22 and 52 mins.

Also upgrade and install patches added so to ensure that every new installation or upgrade to this version will automatically update the randomized offset.
## Why it's being changed

Running all crons for all merchants at the same time puts bursts of pressure on Engagement Cloud, and we're trying to spread the impact.

## How to review / test this change

- Go to developer settings and select cron time frequency
- Make sure that after saving/refreshing the values of the dropdown are correct
- Check carefully the core_config_data and make sure that the saved value for crons follows the pattern (1) AND X<Y
- Confirm that the offset works for the randomly generated offset.

**Upgrade**
- Make sure that previously old type cron values will modified on upgrade
- Modify the crons for different website levels and make sure that the upgrade adds the offset to all of them

**New installation**
- Make sure on a new installation that cron values with offset will be initialised in the table

## Notes

- This only affects crons where the interval is editable via the Developer section.
- You can use this tool to validate also https://crontab.guru/
- For hourly frequency we don't apply the `x-59/y * * * * `, we just set up a random x number from pattern `x * * * * `

Related work items: #132679