Merged PR 30006: Give option to select which tables to migrate
## What's being changed

In this PR we modified our migrate command so our merchants can specify which tables they want to migrate.
So now we can place an option with the table name in migrate command to specify the table we want to migrate.

## Why it's being changed

Currently our migrate command migrates ALL the dotdigital tables, and some merchants find this cumbersome. We often see cases where some subset of data wants to be refreshed (e.g. orders) but this should not reset data in other tables (contacts, catalog) where the re-sync would take time AND possibly lose data (e.g. guests in the contact table).

## How to review / test this change

- run php bin/magento dotdigital:migrate and confirm that all the tables are migrated
- run php bin/magento dotdigital:migrate --table=email_xxxxxxx and confirm that the selected table is migrated
- select an invalid table (a table that doesn't exists in TableProvider acceptance tables and confirm that relevant error message will be printed)

Related work items: #124468