Merged PR 22574: Improved contact import report handling
## What's being changed

The main point of this PR is to add extra conditions before deciding to resubscribe contacts, if they are presented in the list of contact import report faults.

## Why it's being changed

At present we check each contact's `last_subscribed_at` field to see if they have subscribed more recently than the last contact sync - if they have we resubscribe them, if they haven't we unsubscribe them.

We're extending this to check that the contact is also _subscribed_, and _not suppressed_.

In the process of adding these checks we've refactored the processing of the contact import report into its own class, and rewritten the code for clarity.

This PR also eliminates the impact of bug id 119424, where `last_subscribed_at` can be updated at unsubscribe in some versions of Magento. This bug is replicable in 2.3.x, but not in 2.4.x.

## How to review / test this change

- Use this to override `$report`:
```
$report = 'Reason,email,firstname,fullname,gender,lastname,postcode
Suppressed by you,signup@emailsim.io,Chaz,Kangeroo,Male,Kangeroo,123456
Suppressed by you,chaz1@strangerpixel.com,Chaz,Kangeroo,Male,Kangeroo,123456';
```
[When testing locally, the above list should have some email addresses that match contacts in your `email_contact` table, others that don't (i.e. these could be EC-only suppressions). Select contacts with different states e.g. recently subscribed, suppressed, recently subscribed AND suppressed etc.]
- Run importer sync
- Ensure that only contacts that are subscribed, not suppressed AND have last subscribed date between now and the last contact sync time, are resubscribed
- All other contacts should be unsubscribed

## Notes

To retest and cross-check:
- https://github.com/dotmailer/dotmailer-magento2-extension/commit/a44c58c6173b3bb1826f98de200368c94b053bc8
- https://github.com/dotmailer/dotmailer-magento2-extension/commit/56c055d9e98508380f3437581545013e1b752dea

Related work items: #119424, #119452