Merged PR 30197: Filter subscriber automation check by website id
## What's being changed

In this PR we added website id as an extra filter to check if a subscriber email enrolled to subscription automation in email_automation table

## Why it's being changed

If Need to Confirm is switched on, we run a check that says - "If you are confirming your subscription, and you were already enrolled on the subscriber automation, skip". We've now added a website filter to allow the same email address to be enrolled onto subscriber automations on different websites.

## How to review / test this change

- Enable subscriber automation for multiple websites
- Switch on Stores > Configuration > Customers > Newsletter > Subscription Options > Need to confirm
- Subscribe as a guest to website 1
- Use the same email address to subscribe to website 2
- Confirm that the same email address is enrolled to subscriber automation in email_automation table for both websites

Related work items: #145882